[refactor] Fodler를 entity 패키지로 이동 #11
