+++ src/main/kotlin/dev/codeswamp/domain/article/infrastructure/persistence/entity/ArticleMetadataEntity.kt
@@ -1,7 +1,7 @@
 package dev.codeswamp.domain.article.infrastructure.persistence.entity
 
 import dev.codeswamp.domain.comment.entity.Comment
-import dev.codeswamp.domain.folder.domain.Folder
+import dev.codeswamp.domain.folder.domain.entity.Folder
 import dev.codeswamp.domain.user.entity.User
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
