+++ src/main/kotlin/dev/codeswamp/domain/folder/domain/entity/Folder.kt
@@ -1,14 +1,4 @@
-package dev.codeswamp.domain.folder.domain
-
-import dev.codeswamp.domain.article.domain.model.Article
-import dev.codeswamp.domain.user.entity.User
-import jakarta.persistence.Entity
-import jakarta.persistence.FetchType
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.GenerationType
-import jakarta.persistence.Id
-import jakarta.persistence.JoinColumn
-import jakarta.persistence.ManyToOne
+package dev.codeswamp.domain.folder.domain.entity
 
 data class Folder(
     val id: Long? = null,
